% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_imputation.R
\name{boundary}
\alias{boundary}
\title{Correlation Boundary}
\usage{
boundary(corMat, var1, var2, method = "default")
}
\arguments{
\item{corMat}{A correlation matrix}

\item{var1}{A number corresponding to the position in the matrix of the first variable of interest}

\item{var2}{A number corresponding to the position in the matrix of the second variable of interest}

\item{method}{"default" computes the bound based of using values of 0 and pi in the correlative angle matrix. "greek" uses a closed form equation to compute.}
}
\value{
A vector in which the first element is the minimum correlation and the second element is the maximum correlation
}
\description{
Computes the minimum and maximum values for a correlation between two variables in a correlation matrix
}
\examples{
\dontrun{
x <- matrix(c(1,.5,.5,0,.5,1,.5,.3,.5,.5,1,-.1,0,.3,-.1,1), nrow = 4, ncol = 4)
x[4,3] <- NA
x[3,4] <- NA
boundary(x, 4, 3)
}
}
