% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_imputation.R
\name{check_correlation_matrix}
\alias{check_correlation_matrix}
\title{Apply multiple checks to a correlation matrix}
\usage{
check_correlation_matrix(data_matrix)
}
\description{
Apply multiple checks to a correlation matrix
}
