% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_imputation.R
\name{corImpute}
\alias{corImpute}
\title{Impute missing correlation from a matrix}
\usage{
corImpute(corMat, method = "average", interval_prob = NA, interval = NA)
}
\description{
Impute missing correlation from a matrix
}
