% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_consistency_trio.R
\name{correlation_consistency_trio}
\alias{correlation_consistency_trio}
\title{correlation_consistency_trio}
\usage{
correlation_consistency_trio(rAB, rBC, rAC, digits = 2)
}
\arguments{
\item{rAB}{Correlation between A and B}

\item{rBC}{Correlation between B and C}

\item{rAC}{Correlation between A and C}

\item{digits}{The number of digits to round the output to}
}
\description{
Computes the minimum and maximum values for each of three correlations given the other two correlations
}
\examples{
\dontrun{
correlation_consistency_trio(rAB = 0.5, rBC = -0.1, rAC = 0.1)
}
}
