% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_imputation.R
\name{corSwap}
\alias{corSwap}
\title{Swaps columns and rows i and j in a correlation matrix}
\usage{
corSwap(corMat, i, j)

corSwap(corMat, i, j)
}
\arguments{
\item{corMat}{A square correlation matrix}

\item{i, }{j Column and row indices to swap}
}
\value{
The correlation matrix with swapped columns and rows
}
\description{
Swaps columns and rows i and j in a correlation matrix

corSwap
}
