% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_imputation.R
\name{corToAng}
\alias{corToAng}
\title{Correlation Matrix to Correlative Angle Matrix}
\usage{
corToAng(corMat)
}
\arguments{
\item{corMat}{A matrix which contains the correlations between variables}
}
\value{
A correlative angle matrix corresponding to the input correlation matrix
}
\description{
Transform a correlation matrix to its correlative angle matrix form
}
\examples{
\dontrun{
x <- matrix(c(1,.5,.5,0,.5,1,.5,.3,.5,.5,1,-.1,0,.3,-.1,1), nrow = 4, ncol = 4)
x[4,3] <- NA
x[3,4] <- NA
y <- corToAng(x)
y[4,3] <- 0
round(angleToCor(y), 2)
}
}
