% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port.r
\name{define_cutoff}
\alias{define_cutoff}
\title{Violation definition}
\usage{
define_cutoff(
  subgrp,
  var,
  type_var,
  data,
  pruning,
  type_A = "b",
  mediation = FALSE,
  beta,
  A,
  check.side = "both"
)
}
\arguments{
\item{subgrp}{the subgroup produced in port()}

\item{var}{Variables names defining the subgroup}

\item{type_var}{Type of the variable (continuous or categorical)}

\item{data}{Dataset}

\item{pruning}{Boolean, should we remove subgroup like >30 & <40?}

\item{type_A}{Type of the exposure (binary, continuous, or categorical)}

\item{mediation}{Boolean, are we in mediation analysis? Only for table labeling}

\item{beta}{Beta hyperparameter}

\item{A}{Column name of the exposure}

\item{check.side}{Which side of the positivity assumption should be checked? P(A|C), beta is 'zero', P(A|C)>1-beta is 'one', both is 'both' (default)}
}
\value{
Table with the violations (subgroup name | A=1 probability | A level | subgroup size (N) | subgroup size (\%))
}
\description{
Produce a row of results from a problematic subgroup for the user output (table of results).
}
