% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port.r
\name{deport}
\alias{deport}
\title{Decomposition Positivity Regression Trees}
\usage{
deport(
  A,
  M,
  L = NULL,
  effect = "ne",
  type_A = "b",
  type_M = "b",
  type_L = NULL,
  cov.quanti,
  cov.quali,
  data,
  alpha = 0.05,
  beta = "gruber",
  gamma = 2,
  minbucket = 6,
  minsplit = 20,
  maxdepth = 30,
  pruning = FALSE
)
}
\arguments{
\item{A}{Label of the column related to the exposure}

\item{M}{Label of the column related to the mediator (must be binary for effect = "cde")}

\item{L}{Label of the column related to the post-exposure confounder (only one post-treatment confounder is currently allowed).}

\item{effect}{Target estimand for which positivity should be checked ("cde" for controlled direct effect, "ne" for natural effects, or "ie" for interventional effects).}

\item{type_A}{Type of the exposure ('b' for binary, 'c' for continuous, or 'n' for nominal).}

\item{type_M}{Type of the mediator ('b' for binary, 'c' for continuous, or 'n' for nominal).}

\item{type_L}{Type of the post-exposure counfounder ('b' for binary, 'c' for continuous, or 'n' for nominal).}

\item{cov.quanti}{Columns' labels of the quantitative variables in the adjustment set.}

\item{cov.quali}{Columns' labels of the qualitative variables in the adjustment set.}

\item{data}{Dataset, must be a data.frame object.}

\item{alpha}{Subgroup minimal size (as a proportion of the whole sample).}

\item{beta}{Threshold for non-positivity (i.e., extreme exposure's probability).}

\item{gamma}{Maximal number of variables to define a subgroup.}

\item{minbucket}{Rpart hyperparameter, minimum number of individual in the leaves.}

\item{minsplit}{Rpart hyperparameter, minimum number of individual in a node to be split.}

\item{maxdepth}{Rpart hyperparameter, maximum number of successive nodes.}

\item{pruning}{Boolean, should we keep the 'internal' violation (e.g., 20<cov<30).}
}
\value{
Return two lists. The first is related to the exposure positivity and the second is related to the mediator positivity. If effect != "cde", the mediator positivity violations are presented for each counterfactual outcome.
}
\description{
Check the mediation positivity assumptions to identify problematic subgroups/covariables related to either the treatment or the mediator.
This function is a wrapper around port() and works for CDE, NE and IE.
}
