% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port.r
\name{extract_prevalence_subgrp}
\alias{extract_prevalence_subgrp}
\title{Prevalence checking}
\usage{
extract_prevalence_subgrp(x, data, M, L = NULL)
}
\arguments{
\item{x}{Subgroup from a port object (cell in the first column, characters string)}

\item{data}{Other subgroup in which one want to compute the prevalence (data.frame)}

\item{M}{Label of the column related to the mediator}

\item{L}{Label of the column related to the post-treatment confounder}
}
\value{
Return the prevalence of each modality of M in the other treatment group
}
\description{
Check the prevalence of the mediator value for a specific subgroup violating positivity but in the other treatment group.
}
