% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port.r
\name{labels.rpart}
\alias{labels.rpart}
\title{To obtain the cutoff / class}
\usage{
\method{labels}{rpart}(object, data, digits = 4, minlength = 1L, pretty, collapse = TRUE, ...)
}
\arguments{
\item{object}{rpart object}

\item{data}{data.frame}

\item{digits}{miscellaneous}

\item{minlength}{miscellaneous}

\item{pretty}{miscellaneous}

\item{collapse}{miscellaneous}

\item{...}{other arguments to pass}
}
\value{
return the rpart object with the cutoffs (< or >) included
}
\description{
To obtain the cutoff / class
}
