% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_imputation.R
\name{matSolve}
\alias{matSolve}
\title{Solves for the boundary of one missing correlation}
\usage{
matSolve(corMat)
}
\arguments{
\item{corMat}{A correlation matrix with one missing correlation}
}
\value{
A vector in which the first element is the minimum correlation and the second element is the maximum correlation
}
\description{
Automatically detects and returns the boundary of one missing correlation in a correlation matrix
}
\examples{
\dontrun{
x <- matrix(c(1,.5,.5,0,.5,1,.5,.3,.5,.5,1,-.1,0,.3,-.1,1), nrow = 4, ncol = 4)
x[4,3] <- NA
x[3,4] <- NA
matSolve(x)
}
}
