% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{PORT}
\alias{PORT}
\title{PORT: Positive-Definiteness Of (Pearson's) r Tables}
\description{
Check correlation tables for inconsistencies and attempt to isolate which specific correlation estimates may be causing the inconsistency.
}
\section{Important}{

Package is in early stage development.
}

