% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_imputation.R
\name{positive_definite}
\alias{positive_definite}
\title{Check if correlation matrix is positive definite}
\usage{
positive_definite(correlation_matrix)
}
\description{
Check if correlation matrix is positive definite
}
