% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port.r
\name{sport}
\alias{sport}
\title{Sequential positivity Regression Trees}
\usage{
sport(
  A,
  D.bar,
  static = TRUE,
  monotony = FALSE,
  add.subset = NULL,
  time = NULL,
  lag = 0,
  type_A = "b",
  cov.quanti,
  cov.quali,
  data,
  pooling = FALSE,
  alpha = 0.05,
  beta = "gruber",
  gamma = 2,
  minbucket = 6,
  minsplit = 20,
  maxdepth = 30,
  pruning = FALSE
)
}
\arguments{
\item{A}{Intervention indicators. Must be a list of each intervention columns' labels.}

\item{D.bar}{Treatement strategy history's column names. Must be a matrix (n*times).}

\item{static}{Boolean, is the rule static (TRUE, default) or dynamic (FALSE).}

\item{monotony}{Boolean, is the treatment pattern is monotone in data? FALSE by default. See Details.}

\item{add.subset}{Column name related to an indicator of additionnal subseting (1 to keep, 0 to remove), See Details.}

\item{time}{Column name for the time when pooled=T, vector of time when pooling=F (optional).}

\item{lag}{Lagged values to consider for covariates (default is 1). See Details}

\item{type_A}{Type of the intervention ('b' for binary, 'c' for continuous, or 'n' for nominal)}

\item{cov.quanti}{Columns' labels of the quantitative variables in the adjustment set. Must be a list with the same length than A if pooled=FALSE. See Details}

\item{cov.quali}{Columns' labels of the qualitative variables in the adjustment set. Must be a list with the same length than A if pooled=FALSE. See Details}

\item{data}{Dataset, either in the wide format (if pooled=FALSE) or long format (if pooled=TRUE)}

\item{pooling}{Boolean, should the positivity be checked at each time (FALSE, default) or globally (TRUE)?}

\item{alpha}{Subgroup minimal size (as a proportion of the whole sample)}

\item{beta}{Threshold for non-positivity (i.e., extreme exposure's probability).}

\item{gamma}{Maximal number of variables to define a subgroup}

\item{minbucket}{Rpart hyperparameter, minimum number of individual in the leaves}

\item{minsplit}{Rpart hyperparameter, minimum number of individual in a node to be split}

\item{maxdepth}{Rpart hyperparameter, maximum number of successive nodes}

\item{pruning}{Boolean, should we keep the 'internal' violation (e.g., 20<cov<30)}
}
\value{
List of regimen-specific (and time-specific if pooled=FALSE) data.frames with the potential positivity violations identified.
}
\description{
Check the sequential positivity assumption for longitudinal/panel data to identify problematic subgroups/covariables over time.
This function is a wrapper around port() and deals with wide or long data.frames.
}
