% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_imputation.R
\name{thetaSolve}
\alias{thetaSolve}
\title{thetaSolve}
\usage{
thetaSolve(bMat, i, j, partAng = NA)
}
\description{
Solves for theta using logic on NUMPACHAROEN (2013) page 405. Partial angle matrix can be provided to speed up computation (so code can access what it already has calculated previously instead of having to compute everything from the beginning column again).
}
