% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_imputation.R
\name{triangle_to_cor_matrix}
\alias{triangle_to_cor_matrix}
\title{Convert an upper or lower triangle to a full matrix}
\usage{
triangle_to_cor_matrix(triangle)
}
\description{
Convert an upper or lower triangle to a full matrix
}
