% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_imputation.R
\name{within_bounds}
\alias{within_bounds}
\title{Check if all elements of a correlation matrix are within the bounds \link{-1, 1}}
\usage{
within_bounds(correlation_matrix)
}
\description{
Check if all elements of a correlation matrix are within the bounds \link{-1, 1}
}
