% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port.r
\name{port}
\alias{port}
\title{Positivity Regression trees}
\usage{
port(
  A,
  type_A = "b",
  cov.quanti,
  cov.quali,
  data,
  alpha = 0.05,
  beta = "gruber",
  gamma = 2,
  check.side = "both",
  mediation = FALSE,
  graph = "none",
  pruning = FALSE,
  minbucket = 6,
  minsplit = 20,
  maxdepth = 30,
  tweak = 1
)
}
\arguments{
\item{A}{Label of the column related to the exposure}

\item{type_A}{Type of the exposure ('b' for binary, 'c' for continuous, or 'n' for nominal)}

\item{cov.quanti}{Columns' labels of the quantitative variables in the adjustment set}

\item{cov.quali}{Columns' labels of the qualitative variables in the adjustment set}

\item{data}{Dataset, must be a data.frame object}

\item{alpha}{Subgroup minimal size (as a proportion of the whole sample)}

\item{beta}{Threshold for non-positivity (i.e., extreme exposure's probability).}

\item{gamma}{Maximal number of variables to define a subgroup}

\item{check.side}{Which side of the positivity assumption should be checked? P(A|C)<beta is 'zero', P(A|C)>1-beta is 'one', both is 'both' (default)}

\item{mediation}{Boolean, is the exposure a mediator?}

\item{graph}{Provide the trees as additional output?}

\item{pruning}{Boolean, should we keep the 'internal' violation (e.g., >30 & <40)}

\item{minbucket}{Rpart hyperparameter, minimum number of individual in the leaves}

\item{minsplit}{Rpart hyperparameter, minimum number of individual in a node to be split}

\item{maxdepth}{Rpart hyperparameter, maximum number of successive nodes}

\item{tweak}{Text size for the graphs}
}
\value{
Data.frame with all positivity violations (subgroup name | A=1 probability | A level | subgroup size (N) | subgroup relative size (\%))
}
\description{
Check the positivity assumption and identify problematic subgroups/covariables.
}
